Added update link to customer list view
In order to get the customer list update link to work we had to create
a new controller mapping to re-post the existing customer data from the
list to the existing list-customers template. Data pre-population was
taken care of for us automatically by Spring MVC. However, the customer
ID context will be lost and needs to be fixed.