Removed 2 arg customer constructor and replaced with 3 arg
This allows the email address to be used in the construct, which really
should be required. I did not add any validation on the field itself to
use regular expressions, but I did put a hibernate validation on it to
ensure that the value is unique in the DB when inserted or updated.

Functionality to validate the email address pattern will be added later,
although until the email can actually be verified it's not exactly the
most helpful validation.